@@ -76,7 +76,8 @@ target-version = ['py39']
 
 [tool.ruff]
 line-length = 120
-target-version = 'py37'
+target-version = 'py38'
+
 
 [tool.ruff.lint]
 ignore = ["E501","E731"]
